ARTS-48 Rename component and some style tweaks
