+++ src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { AssignedSitesPageComponent } from './administrable-sites-page/assigned-sites-page.component';
+import { AssignedSitesPageComponent } from './assigned-sites-page/assigned-sites-page.component';
 import { LandingPageComponent } from './landing-page/landing-page.component';
 import {
   AuthGuardService as AuthGuard
