+++ src/app/app.component.sass
@@ -1,0 +1,5 @@
+$primary-color: #333
+
+.logo 
+    height: 45px
+    width: 150px
