+++ src/app/assigned-sites-page/assigned-sites-page.component.html
@@ -1,4 +1,4 @@
-<div class="container">
+<div class="container mt-4">
     <div class="title">
         <h2>Assigned Sites</h2>
     </div>
