+++ src/app/assigned-sites-page/assigned-sites-page.component.spec.ts
@@ -1,14 +1,19 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { SiteService } from '../services/site.service';
+import { MockSiteService } from '../services/site.service.mock';
 
 import { AssignedSitesPageComponent } from './assigned-sites-page.component';
 
-describe('AssignedSitesComponent', () => {
+describe('AssignedSitesPageComponent', () => {
   let component: AssignedSitesPageComponent;
   let fixture: ComponentFixture<AssignedSitesPageComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ AssignedSitesPageComponent ]
+      declarations: [ AssignedSitesPageComponent ],
+      providers: [
+        { provide: SiteService, useClass: MockSiteService }
+      ]
     })
     .compileComponents();
   });
