+++ src/app/assigned-sites-page/assigned-sites-page.component.ts
